perf: pack cached_value_size as u32 (#285)
cached_value_size was usize (8 bytes on 64-bit) but the maximum value
size is 512 MB (MAX_VALUE_LEN), well within u32 range (~4 GB).

packing it as u32 shrinks Entry from 44 to 40 bytes. combined with
the version side table from the previous PR, Entry is now 12 bytes
smaller than the original 52-byte layout. ENTRY_OVERHEAD reduced
from 120 to 116.

+++ crates/ember-core/src/keyspace/hash.rs
@@ -203,7 +203,7 @@ impl Keyspace {
         entry.touch();
 
         let new_value_size = memory::value_size(&entry.value);
-        entry.cached_value_size = new_value_size;
+        entry.cached_value_size = new_value_size as u32;
         let new_entry_size = key.len() + new_value_size + memory::ENTRY_OVERHEAD;
         self.memory.adjust(old_entry_size, new_entry_size);
         self.bump_version(key);
